+++ service/policy/db/query.sql.go
@@ -448,16 +448,13 @@ LEFT JOIN (
         k.definition_id,
         JSONB_AGG(
             DISTINCT JSONB_BUILD_OBJECT(
-                'key', JSONB_BUILD_OBJECT(
-                    'id', kask.id,
-                    'key_id', kask.key_id,
-                    'key_status', kask.key_status,
-                    'key_mode', kask.key_mode,
-                    'key_algorithm', kask.key_algorithm,
-                    'public_key_ctx', kask.public_key_ctx
-                ),
-                'kas_id', kask.key_access_server_id,
-                'kas_uri', kas.uri
+                'kas_uri', kas.uri,
+                'kas_id', kas.id,
+                'public_key', JSONB_BUILD_OBJECT(
+                     'algorithm', kask.key_algorithm::INTEGER,
+                     'kid', kask.key_id,
+                     'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
+                )
             )
         ) FILTER (WHERE kask.id IS NOT NULL) AS keys
     FROM attribute_definition_public_key_map k
@@ -539,16 +536,13 @@ type GetAttributeRow struct {
 //	        k.definition_id,
 //	        JSONB_AGG(
 //	            DISTINCT JSONB_BUILD_OBJECT(
-//	                'key', JSONB_BUILD_OBJECT(
-//	                    'id', kask.id,
-//	                    'key_id', kask.key_id,
-//	                    'key_status', kask.key_status,
-//	                    'key_mode', kask.key_mode,
-//	                    'key_algorithm', kask.key_algorithm,
-//	                    'public_key_ctx', kask.public_key_ctx
-//	                ),
-//	                'kas_id', kask.key_access_server_id,
-//	                'kas_uri', kas.uri
+//	                'kas_uri', kas.uri,
+//	                'kas_id', kas.id,
+//	                'public_key', JSONB_BUILD_OBJECT(
+//	                     'algorithm', kask.key_algorithm::INTEGER,
+//	                     'kid', kask.key_id,
+//	                     'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
+//	                )
 //	            )
 //	        ) FILTER (WHERE kask.id IS NOT NULL) AS keys
 //	    FROM attribute_definition_public_key_map k
@@ -604,15 +598,12 @@ LEFT JOIN (
         k.value_id,
         JSONB_AGG(
             DISTINCT JSONB_BUILD_OBJECT(
-                'kas_id', kask.key_access_server_id,
                 'kas_uri', kas.uri,
-                'key', JSONB_BUILD_OBJECT(
-                    'id', kask.id,
-                    'key_id', kask.key_id,
-                    'key_status', kask.key_status,
-                    'key_mode', kask.key_mode,
-                    'key_algorithm', kask.key_algorithm,
-                    'public_key_ctx', kask.public_key_ctx
+                'kas_id', kas.id,
+                'public_key', JSONB_BUILD_OBJECT(
+                     'algorithm', kask.key_algorithm::INTEGER,
+                     'kid', kask.key_id,
+                     'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
                 )
             )
         ) FILTER (WHERE kask.id IS NOT NULL) AS keys
@@ -669,15 +660,12 @@ type GetAttributeValueRow struct {
 //	        k.value_id,
 //	        JSONB_AGG(
 //	            DISTINCT JSONB_BUILD_OBJECT(
-//	                'kas_id', kask.key_access_server_id,
 //	                'kas_uri', kas.uri,
-//	                'key', JSONB_BUILD_OBJECT(
-//	                    'id', kask.id,
-//	                    'key_id', kask.key_id,
-//	                    'key_status', kask.key_status,
-//	                    'key_mode', kask.key_mode,
-//	                    'key_algorithm', kask.key_algorithm,
-//	                    'public_key_ctx', kask.public_key_ctx
+//	                'kas_id', kas.id,
+//	                'public_key', JSONB_BUILD_OBJECT(
+//	                     'algorithm', kask.key_algorithm::INTEGER,
+//	                     'kid', kask.key_id,
+//	                     'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
 //	                )
 //	            )
 //	        ) FILTER (WHERE kask.id IS NOT NULL) AS keys
@@ -726,18 +714,17 @@ LEFT JOIN (
             kask.key_access_server_id,
             JSONB_AGG(
                 DISTINCT JSONB_BUILD_OBJECT(
-                    'kas_id', kask.key_access_server_id,
-                    'key', JSONB_BUILD_OBJECT(
-                        'id', kask.id,
-                        'key_id', kask.key_id,
-                        'key_status', kask.key_status,
-                        'key_mode', kask.key_mode,
-                        'key_algorithm', kask.key_algorithm,
-                        'public_key_ctx', kask.public_key_ctx
+                    'kas_uri', kas.uri,
+                    'kas_id', kas.id,
+                    'public_key', JSONB_BUILD_OBJECT(
+                         'algorithm', kask.key_algorithm::INTEGER,
+                         'kid', kask.key_id,
+                         'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
                     )
                 )
             ) FILTER (WHERE kask.id IS NOT NULL) AS keys
         FROM key_access_server_keys kask
+        INNER JOIN key_access_servers kas ON kask.key_access_server_id = kas.id
         GROUP BY kask.key_access_server_id
     ) kask_keys ON kas.id = kask_keys.key_access_server_id
 WHERE ($1::uuid IS NULL OR kas.id = $1::uuid)
@@ -783,18 +770,17 @@ type GetKeyAccessServerRow struct {
 //	            kask.key_access_server_id,
 //	            JSONB_AGG(
 //	                DISTINCT JSONB_BUILD_OBJECT(
-//	                    'kas_id', kask.key_access_server_id,
-//	                    'key', JSONB_BUILD_OBJECT(
-//	                        'id', kask.id,
-//	                        'key_id', kask.key_id,
-//	                        'key_status', kask.key_status,
-//	                        'key_mode', kask.key_mode,
-//	                        'key_algorithm', kask.key_algorithm,
-//	                        'public_key_ctx', kask.public_key_ctx
+//	                    'kas_uri', kas.uri,
+//	                    'kas_id', kas.id,
+//	                    'public_key', JSONB_BUILD_OBJECT(
+//	                         'algorithm', kask.key_algorithm::INTEGER,
+//	                         'kid', kask.key_id,
+//	                         'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
 //	                    )
 //	                )
 //	            ) FILTER (WHERE kask.id IS NOT NULL) AS keys
 //	        FROM key_access_server_keys kask
+//	        INNER JOIN key_access_servers kas ON kask.key_access_server_id = kas.id
 //	        GROUP BY kask.key_access_server_id
 //	    ) kask_keys ON kas.id = kask_keys.key_access_server_id
 //	WHERE ($1::uuid IS NULL OR kas.id = $1::uuid)
@@ -838,15 +824,12 @@ LEFT JOIN (
         k.namespace_id,
         JSONB_AGG(
             DISTINCT JSONB_BUILD_OBJECT(
-                'kas_id', kask.key_access_server_id,
                 'kas_uri', kas.uri,
-                'key', JSONB_BUILD_OBJECT(
-                    'id', kask.id,
-                    'key_id', kask.key_id,
-                    'key_status', kask.key_status,
-                    'key_mode', kask.key_mode,
-                    'key_algorithm', kask.key_algorithm,
-                    'public_key_ctx', kask.public_key_ctx
+                'kas_id', kas.id,
+                'public_key', JSONB_BUILD_OBJECT(
+                     'algorithm', kask.key_algorithm::INTEGER,
+                     'kid', kask.key_id,
+                     'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
                 )
             )
         ) FILTER (WHERE kask.id IS NOT NULL) AS keys
@@ -900,15 +883,12 @@ type GetNamespaceRow struct {
 //	        k.namespace_id,
 //	        JSONB_AGG(
 //	            DISTINCT JSONB_BUILD_OBJECT(
-//	                'kas_id', kask.key_access_server_id,
 //	                'kas_uri', kas.uri,
-//	                'key', JSONB_BUILD_OBJECT(
-//	                    'id', kask.id,
-//	                    'key_id', kask.key_id,
-//	                    'key_status', kask.key_status,
-//	                    'key_mode', kask.key_mode,
-//	                    'key_algorithm', kask.key_algorithm,
-//	                    'public_key_ctx', kask.public_key_ctx
+//	                'kas_id', kas.id,
+//	                'public_key', JSONB_BUILD_OBJECT(
+//	                     'algorithm', kask.key_algorithm::INTEGER,
+//	                     'kid', kask.key_id,
+//	                     'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
 //	                )
 //	            )
 //	        ) FILTER (WHERE kask.id IS NOT NULL) AS keys
@@ -1617,18 +1597,17 @@ LEFT JOIN (
             kask.key_access_server_id,
             JSONB_AGG(
                 DISTINCT JSONB_BUILD_OBJECT(
-                    'kas_id', kask.key_access_server_id,
-                    'key', JSONB_BUILD_OBJECT(
-                        'id', kask.id,
-                        'key_id', kask.key_id,
-                        'key_status', kask.key_status,
-                        'key_mode', kask.key_mode,
-                        'key_algorithm', kask.key_algorithm,
-                        'public_key_ctx', kask.public_key_ctx
+                    'kas_uri', kas.uri,
+                    'kas_id', kas.id,
+                    'public_key', JSONB_BUILD_OBJECT(
+                         'algorithm', kask.key_algorithm::INTEGER,
+                         'kid', kask.key_id,
+                         'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
                     )
                 )
             ) FILTER (WHERE kask.id IS NOT NULL) AS keys
         FROM key_access_server_keys kask
+        INNER JOIN key_access_servers kas ON kask.key_access_server_id = kas.id
         GROUP BY kask.key_access_server_id
     ) kask_keys ON kas.id = kask_keys.key_access_server_id
 LIMIT $2 
@@ -1672,18 +1651,17 @@ type ListKeyAccessServersRow struct {
 //	            kask.key_access_server_id,
 //	            JSONB_AGG(
 //	                DISTINCT JSONB_BUILD_OBJECT(
-//	                    'kas_id', kask.key_access_server_id,
-//	                    'key', JSONB_BUILD_OBJECT(
-//	                        'id', kask.id,
-//	                        'key_id', kask.key_id,
-//	                        'key_status', kask.key_status,
-//	                        'key_mode', kask.key_mode,
-//	                        'key_algorithm', kask.key_algorithm,
-//	                        'public_key_ctx', kask.public_key_ctx
+//	                    'kas_uri', kas.uri,
+//	                    'kas_id', kas.id,
+//	                    'public_key', JSONB_BUILD_OBJECT(
+//	                         'algorithm', kask.key_algorithm::INTEGER,
+//	                         'kid', kask.key_id,
+//	                         'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
 //	                    )
 //	                )
 //	            ) FILTER (WHERE kask.id IS NOT NULL) AS keys
 //	        FROM key_access_server_keys kask
+//	        INNER JOIN key_access_servers kas ON kask.key_access_server_id = kas.id
 //	        GROUP BY kask.key_access_server_id
 //	    ) kask_keys ON kas.id = kask_keys.key_access_server_id
 //	LIMIT $2
@@ -3430,10 +3408,11 @@ const getBaseKey = `-- name: getBaseKey :one
 SELECT
     DISTINCT JSONB_BUILD_OBJECT(
        'kas_uri', kas.uri,
+       'kas_id', kas.id,
        'public_key', JSONB_BUILD_OBJECT(
-            'algorithm', kask.key_algorithm::TEXT,
+            'algorithm', kask.key_algorithm::INTEGER,
             'kid', kask.key_id,
-            'pem', kask.public_key_ctx ->> 'pem'
+            'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
        )
     ) AS base_keys
 FROM base_keys bk
@@ -3448,10 +3427,11 @@ INNER JOIN key_access_servers kas ON kask.key_access_server_id = kas.id
 //	SELECT
 //	    DISTINCT JSONB_BUILD_OBJECT(
 //	       'kas_uri', kas.uri,
+//	       'kas_id', kas.id,
 //	       'public_key', JSONB_BUILD_OBJECT(
-//	            'algorithm', kask.key_algorithm::TEXT,
+//	            'algorithm', kask.key_algorithm::INTEGER,
 //	            'kid', kask.key_id,
-//	            'pem', kask.public_key_ctx ->> 'pem'
+//	            'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
 //	       )
 //	    ) AS base_keys
 //	FROM base_keys bk
@@ -3966,15 +3946,12 @@ WITH target_definition AS (
             k.definition_id,
             JSONB_AGG(
                 DISTINCT JSONB_BUILD_OBJECT(
-                    'kas_id', kask.key_access_server_id,
                     'kas_uri', kas.uri,
-                    'key', JSONB_BUILD_OBJECT(
-                        'id', kask.id,
-                        'key_id', kask.key_id,
-                        'key_status', kask.key_status,
-                        'key_mode', kask.key_mode,
-                        'key_algorithm', kask.key_algorithm,
-                        'public_key_ctx', kask.public_key_ctx
+                    'kas_id', kas.id,
+                    'public_key', JSONB_BUILD_OBJECT(
+                         'algorithm', kask.key_algorithm::INTEGER,
+                         'kid', kask.key_id,
+                         'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
                     )
                 )
             ) FILTER (WHERE kask.id IS NOT NULL) AS keys
@@ -4015,15 +3992,12 @@ namespaces AS (
             k.namespace_id,
             JSONB_AGG(
                 DISTINCT JSONB_BUILD_OBJECT(
-                    'kas_id', kask.key_access_server_id,
                     'kas_uri', kas.uri,
-                    'key', JSONB_BUILD_OBJECT(
-                        'id', kask.id,
-                        'key_id', kask.key_id,
-                        'key_status', kask.key_status,
-                        'key_mode', kask.key_mode,
-                        'key_algorithm', kask.key_algorithm,
-                        'public_key_ctx', kask.public_key_ctx
+                    'kas_id', kas.id,
+                    'public_key', JSONB_BUILD_OBJECT(
+                         'algorithm', kask.key_algorithm::INTEGER,
+                         'kid', kask.key_id,
+                         'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
                     )
                 )
             ) FILTER (WHERE kask.id IS NOT NULL) AS keys
@@ -4135,15 +4109,12 @@ values AS (
             k.value_id,
             JSONB_AGG(
                 DISTINCT JSONB_BUILD_OBJECT(
-                    'kas_id', kask.key_access_server_id,
                     'kas_uri', kas.uri,
-                    'key', JSONB_BUILD_OBJECT(
-                        'id', kask.id,
-                        'key_id', kask.key_id,
-                        'key_status', kask.key_status,
-                        'key_mode', kask.key_mode,
-                        'key_algorithm', kask.key_algorithm,
-                        'public_key_ctx', kask.public_key_ctx
+                    'kas_id', kas.id,
+                    'public_key', JSONB_BUILD_OBJECT(
+                         'algorithm', kask.key_algorithm::INTEGER,
+                         'kid', kask.key_id,
+                         'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
                     )
                 )
             ) FILTER (WHERE kask.id IS NOT NULL) AS keys
@@ -4212,15 +4183,12 @@ type listAttributesByDefOrValueFqnsRow struct {
 //	            k.definition_id,
 //	            JSONB_AGG(
 //	                DISTINCT JSONB_BUILD_OBJECT(
-//	                    'kas_id', kask.key_access_server_id,
 //	                    'kas_uri', kas.uri,
-//	                    'key', JSONB_BUILD_OBJECT(
-//	                        'id', kask.id,
-//	                        'key_id', kask.key_id,
-//	                        'key_status', kask.key_status,
-//	                        'key_mode', kask.key_mode,
-//	                        'key_algorithm', kask.key_algorithm,
-//	                        'public_key_ctx', kask.public_key_ctx
+//	                    'kas_id', kas.id,
+//	                    'public_key', JSONB_BUILD_OBJECT(
+//	                         'algorithm', kask.key_algorithm::INTEGER,
+//	                         'kid', kask.key_id,
+//	                         'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
 //	                    )
 //	                )
 //	            ) FILTER (WHERE kask.id IS NOT NULL) AS keys
@@ -4261,15 +4229,12 @@ type listAttributesByDefOrValueFqnsRow struct {
 //	            k.namespace_id,
 //	            JSONB_AGG(
 //	                DISTINCT JSONB_BUILD_OBJECT(
-//	                    'kas_id', kask.key_access_server_id,
 //	                    'kas_uri', kas.uri,
-//	                    'key', JSONB_BUILD_OBJECT(
-//	                        'id', kask.id,
-//	                        'key_id', kask.key_id,
-//	                        'key_status', kask.key_status,
-//	                        'key_mode', kask.key_mode,
-//	                        'key_algorithm', kask.key_algorithm,
-//	                        'public_key_ctx', kask.public_key_ctx
+//	                    'kas_id', kas.id,
+//	                    'public_key', JSONB_BUILD_OBJECT(
+//	                         'algorithm', kask.key_algorithm::INTEGER,
+//	                         'kid', kask.key_id,
+//	                         'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
 //	                    )
 //	                )
 //	            ) FILTER (WHERE kask.id IS NOT NULL) AS keys
@@ -4381,15 +4346,12 @@ type listAttributesByDefOrValueFqnsRow struct {
 //	            k.value_id,
 //	            JSONB_AGG(
 //	                DISTINCT JSONB_BUILD_OBJECT(
-//	                    'kas_id', kask.key_access_server_id,
 //	                    'kas_uri', kas.uri,
-//	                    'key', JSONB_BUILD_OBJECT(
-//	                        'id', kask.id,
-//	                        'key_id', kask.key_id,
-//	                        'key_status', kask.key_status,
-//	                        'key_mode', kask.key_mode,
-//	                        'key_algorithm', kask.key_algorithm,
-//	                        'public_key_ctx', kask.public_key_ctx
+//	                    'kas_id', kas.id,
+//	                    'public_key', JSONB_BUILD_OBJECT(
+//	                         'algorithm', kask.key_algorithm::INTEGER,
+//	                         'kid', kask.key_id,
+//	                         'pem', CONVERT_FROM(DECODE(kask.public_key_ctx ->> 'pem', 'base64'), 'UTF8')
 //	                    )
 //	                )
 //	            ) FILTER (WHERE kask.id IS NOT NULL) AS keys
@@ -4884,24 +4846,28 @@ func (q *Queries) listRegisteredResources(ctx context.Context, arg listRegistere
 
 const listSubjectMappings = `-- name: listSubjectMappings :many
 
-WITH counted AS (
+WITH subject_actions AS (
+    SELECT
+        sma.subject_mapping_id,
+        COALESCE(
+            JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name)) FILTER (WHERE a.is_standard = TRUE),
+            '[]'::JSONB
+        ) AS standard_actions,
+        COALESCE(
+            JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name)) FILTER (WHERE a.is_standard = FALSE),
+            '[]'::JSONB
+        ) AS custom_actions
+    FROM subject_mapping_actions sma
+    JOIN actions a ON sma.action_id = a.id
+    GROUP BY sma.subject_mapping_id
+), counted AS (
     SELECT COUNT(sm.id) AS total
     FROM subject_mappings sm
 )
 SELECT
     sm.id,
-    (
-        SELECT JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name))
-        FROM actions a
-        JOIN subject_mapping_actions sma ON sma.action_id = a.id
-        WHERE sma.subject_mapping_id = sm.id AND a.is_standard = TRUE
-    ) AS standard_actions,
-    (
-        SELECT JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name))
-        FROM actions a
-        JOIN subject_mapping_actions sma ON sma.action_id = a.id
-        WHERE sma.subject_mapping_id = sm.id AND a.is_standard = FALSE
-    ) AS custom_actions,
+    sa.standard_actions,
+    sa.custom_actions,
     JSON_STRIP_NULLS(JSON_BUILD_OBJECT('labels', sm.metadata -> 'labels', 'created_at', sm.created_at, 'updated_at', sm.updated_at)) AS metadata,
     JSON_BUILD_OBJECT(
         'id', scs.id,
@@ -4917,10 +4883,19 @@ SELECT
     counted.total
 FROM subject_mappings sm
 CROSS JOIN counted
+LEFT JOIN subject_actions sa ON sm.id = sa.subject_mapping_id
 LEFT JOIN attribute_values av ON sm.attribute_value_id = av.id
 LEFT JOIN attribute_fqns fqns ON av.id = fqns.value_id
 LEFT JOIN subject_condition_set scs ON scs.id = sm.subject_condition_set_id
-GROUP BY av.id, sm.id, scs.id, counted.total, fqns.fqn
+GROUP BY
+    sm.id,
+    sa.standard_actions,
+    sa.custom_actions,
+    sm.metadata, sm.created_at, sm.updated_at, -- for metadata object
+    scs.id, scs.metadata, scs.created_at, scs.updated_at, scs.condition, -- for subject_condition_set object
+    av.id, av.value, av.active, -- for attribute_value object
+    fqns.fqn,
+    counted.total
 LIMIT $2
 OFFSET $1
 `
@@ -4931,37 +4906,41 @@ type listSubjectMappingsParams struct {
 }
 
 type listSubjectMappingsRow struct {
-	ID                  string `json:"id"`
-	StandardActions     []byte `json:"standard_actions"`
-	CustomActions       []byte `json:"custom_actions"`
-	Metadata            []byte `json:"metadata"`
-	SubjectConditionSet []byte `json:"subject_condition_set"`
-	AttributeValue      []byte `json:"attribute_value"`
-	Total               int64  `json:"total"`
+	ID                  string      `json:"id"`
+	StandardActions     interface{} `json:"standard_actions"`
+	CustomActions       interface{} `json:"custom_actions"`
+	Metadata            []byte      `json:"metadata"`
+	SubjectConditionSet []byte      `json:"subject_condition_set"`
+	AttributeValue      []byte      `json:"attribute_value"`
+	Total               int64       `json:"total"`
 }
 
 // --------------------------------------------------------------
 // SUBJECT MAPPINGS
 // --------------------------------------------------------------
 //
-//	WITH counted AS (
+//	WITH subject_actions AS (
+//	    SELECT
+//	        sma.subject_mapping_id,
+//	        COALESCE(
+//	            JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name)) FILTER (WHERE a.is_standard = TRUE),
+//	            '[]'::JSONB
+//	        ) AS standard_actions,
+//	        COALESCE(
+//	            JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name)) FILTER (WHERE a.is_standard = FALSE),
+//	            '[]'::JSONB
+//	        ) AS custom_actions
+//	    FROM subject_mapping_actions sma
+//	    JOIN actions a ON sma.action_id = a.id
+//	    GROUP BY sma.subject_mapping_id
+//	), counted AS (
 //	    SELECT COUNT(sm.id) AS total
 //	    FROM subject_mappings sm
 //	)
 //	SELECT
 //	    sm.id,
-//	    (
-//	        SELECT JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name))
-//	        FROM actions a
-//	        JOIN subject_mapping_actions sma ON sma.action_id = a.id
-//	        WHERE sma.subject_mapping_id = sm.id AND a.is_standard = TRUE
-//	    ) AS standard_actions,
-//	    (
-//	        SELECT JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name))
-//	        FROM actions a
-//	        JOIN subject_mapping_actions sma ON sma.action_id = a.id
-//	        WHERE sma.subject_mapping_id = sm.id AND a.is_standard = FALSE
-//	    ) AS custom_actions,
+//	    sa.standard_actions,
+//	    sa.custom_actions,
 //	    JSON_STRIP_NULLS(JSON_BUILD_OBJECT('labels', sm.metadata -> 'labels', 'created_at', sm.created_at, 'updated_at', sm.updated_at)) AS metadata,
 //	    JSON_BUILD_OBJECT(
 //	        'id', scs.id,
@@ -4977,10 +4956,19 @@ type listSubjectMappingsRow struct {
 //	    counted.total
 //	FROM subject_mappings sm
 //	CROSS JOIN counted
+//	LEFT JOIN subject_actions sa ON sm.id = sa.subject_mapping_id
 //	LEFT JOIN attribute_values av ON sm.attribute_value_id = av.id
 //	LEFT JOIN attribute_fqns fqns ON av.id = fqns.value_id
 //	LEFT JOIN subject_condition_set scs ON scs.id = sm.subject_condition_set_id
-//	GROUP BY av.id, sm.id, scs.id, counted.total, fqns.fqn
+//	GROUP BY
+//	    sm.id,
+//	    sa.standard_actions,
+//	    sa.custom_actions,
+//	    sm.metadata, sm.created_at, sm.updated_at, -- for metadata object
+//	    scs.id, scs.metadata, scs.created_at, scs.updated_at, scs.condition, -- for subject_condition_set object
+//	    av.id, av.value, av.active, -- for attribute_value object
+//	    fqns.fqn,
+//	    counted.total
 //	LIMIT $2
 //	OFFSET $1
 func (q *Queries) listSubjectMappings(ctx context.Context, arg listSubjectMappingsParams) ([]listSubjectMappingsRow, error) {
@@ -5012,20 +5000,25 @@ func (q *Queries) listSubjectMappings(ctx context.Context, arg listSubjectMappin
 }
 
 const matchSubjectMappings = `-- name: matchSubjectMappings :many
+WITH subject_actions AS (
+    SELECT
+        sma.subject_mapping_id,
+        COALESCE(
+            JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name)) FILTER (WHERE a.is_standard = TRUE),
+            '[]'::JSONB
+        ) AS standard_actions,
+        COALESCE(
+            JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name)) FILTER (WHERE a.is_standard = FALSE),
+            '[]'::JSONB
+        ) AS custom_actions
+    FROM subject_mapping_actions sma
+    JOIN actions a ON sma.action_id = a.id
+    GROUP BY sma.subject_mapping_id
+)
 SELECT
     sm.id,
-    (
-        SELECT JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name))
-        FROM actions a
-        JOIN subject_mapping_actions sma ON sma.action_id = a.id
-        WHERE sma.subject_mapping_id = sm.id AND a.is_standard = TRUE
-    ) AS standard_actions,
-    (
-        SELECT JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name))
-        FROM actions a
-        JOIN subject_mapping_actions sma ON sma.action_id = a.id
-        WHERE sma.subject_mapping_id = sm.id AND a.is_standard = FALSE
-    ) AS custom_actions,
+    sa.standard_actions,
+    sa.custom_actions,
     JSON_BUILD_OBJECT(
         'id', scs.id,
         'subject_sets', scs.condition
@@ -5037,40 +5030,54 @@ SELECT
         'fqn', fqns.fqn
     ) AS attribute_value
 FROM subject_mappings sm
+LEFT JOIN subject_actions sa ON sm.id = sa.subject_mapping_id
 LEFT JOIN attribute_values av ON sm.attribute_value_id = av.id
 LEFT JOIN attribute_definitions ad ON av.attribute_definition_id = ad.id
 LEFT JOIN attribute_namespaces ns ON ad.namespace_id = ns.id
 LEFT JOIN attribute_fqns fqns ON av.id = fqns.value_id
 LEFT JOIN subject_condition_set scs ON scs.id = sm.subject_condition_set_id
-WHERE ns.active = true AND ad.active = true and av.active = true
-AND scs.selector_values && $1::TEXT[]
-GROUP BY av.id, sm.id, scs.id, fqns.fqn
+WHERE
+    ns.active = TRUE
+    AND ad.active = TRUE
+    AND av.active = TRUE
+    AND scs.selector_values && $1::TEXT[]
+GROUP BY
+    sm.id,
+    sa.standard_actions,
+    sa.custom_actions,
+    scs.id, scs.condition,
+    av.id, av.value, av.active, fqns.fqn
 `
 
 type matchSubjectMappingsRow struct {
-	ID                  string `json:"id"`
-	StandardActions     []byte `json:"standard_actions"`
-	CustomActions       []byte `json:"custom_actions"`
-	SubjectConditionSet []byte `json:"subject_condition_set"`
-	AttributeValue      []byte `json:"attribute_value"`
+	ID                  string      `json:"id"`
+	StandardActions     interface{} `json:"standard_actions"`
+	CustomActions       interface{} `json:"custom_actions"`
+	SubjectConditionSet []byte      `json:"subject_condition_set"`
+	AttributeValue      []byte      `json:"attribute_value"`
 }
 
 // matchSubjectMappings
 //
+//	WITH subject_actions AS (
+//	    SELECT
+//	        sma.subject_mapping_id,
+//	        COALESCE(
+//	            JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name)) FILTER (WHERE a.is_standard = TRUE),
+//	            '[]'::JSONB
+//	        ) AS standard_actions,
+//	        COALESCE(
+//	            JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name)) FILTER (WHERE a.is_standard = FALSE),
+//	            '[]'::JSONB
+//	        ) AS custom_actions
+//	    FROM subject_mapping_actions sma
+//	    JOIN actions a ON sma.action_id = a.id
+//	    GROUP BY sma.subject_mapping_id
+//	)
 //	SELECT
 //	    sm.id,
-//	    (
-//	        SELECT JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name))
-//	        FROM actions a
-//	        JOIN subject_mapping_actions sma ON sma.action_id = a.id
-//	        WHERE sma.subject_mapping_id = sm.id AND a.is_standard = TRUE
-//	    ) AS standard_actions,
-//	    (
-//	        SELECT JSONB_AGG(JSONB_BUILD_OBJECT('id', a.id, 'name', a.name))
-//	        FROM actions a
-//	        JOIN subject_mapping_actions sma ON sma.action_id = a.id
-//	        WHERE sma.subject_mapping_id = sm.id AND a.is_standard = FALSE
-//	    ) AS custom_actions,
+//	    sa.standard_actions,
+//	    sa.custom_actions,
 //	    JSON_BUILD_OBJECT(
 //	        'id', scs.id,
 //	        'subject_sets', scs.condition
@@ -5082,14 +5089,23 @@ type matchSubjectMappingsRow struct {
 //	        'fqn', fqns.fqn
 //	    ) AS attribute_value
 //	FROM subject_mappings sm
+//	LEFT JOIN subject_actions sa ON sm.id = sa.subject_mapping_id
 //	LEFT JOIN attribute_values av ON sm.attribute_value_id = av.id
 //	LEFT JOIN attribute_definitions ad ON av.attribute_definition_id = ad.id
 //	LEFT JOIN attribute_namespaces ns ON ad.namespace_id = ns.id
 //	LEFT JOIN attribute_fqns fqns ON av.id = fqns.value_id
 //	LEFT JOIN subject_condition_set scs ON scs.id = sm.subject_condition_set_id
-//	WHERE ns.active = true AND ad.active = true and av.active = true
-//	AND scs.selector_values && $1::TEXT[]
-//	GROUP BY av.id, sm.id, scs.id, fqns.fqn
+//	WHERE
+//	    ns.active = TRUE
+//	    AND ad.active = TRUE
+//	    AND av.active = TRUE
+//	    AND scs.selector_values && $1::TEXT[]
+//	GROUP BY
+//	    sm.id,
+//	    sa.standard_actions,
+//	    sa.custom_actions,
+//	    scs.id, scs.condition,
+//	    av.id, av.value, av.active, fqns.fqn
 func (q *Queries) matchSubjectMappings(ctx context.Context, selectors []string) ([]matchSubjectMappingsRow, error) {
 	rows, err := q.db.Query(ctx, matchSubjectMappings, selectors)
 	if err != nil {