+++ protocol/go/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.4.0](https://github.com/opentdf/platform/compare/protocol/go/v0.3.6...protocol/go/v0.4.0) (2025-06-05)
+
+
+### Features
+
+* **authz:** improve v2 request proto validation ([#2357](https://github.com/opentdf/platform/issues/2357)) ([f927b99](https://github.com/opentdf/platform/commit/f927b994149079947cac1d1386f2bfb9a52139a0))
+* **policy:** cache SubjectConditionSet selectors in dedicated column maintained via trigger ([#2320](https://github.com/opentdf/platform/issues/2320)) ([215791f](https://github.com/opentdf/platform/commit/215791f2185d6cacfa4a8ae4a009739ee30bfc66))
+* **policy:** Return Simple Kas Keys from non-Key RPCs ([#2387](https://github.com/opentdf/platform/issues/2387)) ([5113e0e](https://github.com/opentdf/platform/commit/5113e0edbe0260d0937a62932671b40ca5cfcbf4))
+* **policy:** Update simple kas key ([#2378](https://github.com/opentdf/platform/issues/2378)) ([09d8239](https://github.com/opentdf/platform/commit/09d82390a06e22a8787118cd0ec7d97311e85363))
+
+
+### Bug Fixes
+
+* **policy:** protovalidate deprecated action types and removal of gRPC gateway in subject mappings svc ([#2377](https://github.com/opentdf/platform/issues/2377)) ([54a6de0](https://github.com/opentdf/platform/commit/54a6de03d8796b0fe72edc381ce514927bdcd793))
+* **policy:** remove gRPC gateway in policy except where needed ([#2382](https://github.com/opentdf/platform/issues/2382)) ([1937acb](https://github.com/opentdf/platform/commit/1937acb3fff5e6216808ac233d3a34b869901b44))
+
 ## [0.3.6](https://github.com/opentdf/platform/compare/protocol/go/v0.3.5...protocol/go/v0.3.6) (2025-05-27)